+++ packages/generator/src/resources/shortcodes.ts
@@ -155,7 +155,7 @@ export default {
   // 😷 face with medical mask
   '1F637': ['medical_mask', 'mask'],
   // 😵 knocked-out face
-  '1F635': ['dizzy_face'],
+  '1F635': ['dizzy_face', 'knocked_out'],
   // 😲 astonished face
   '1F632': ['astonished_face', 'astonished'],
   // 😳 flushed face
@@ -911,7 +911,7 @@ export default {
   // 🎿 skis
   '1F3BF': ['ski'],
   // 🎯 bullseye
-  '1F3AF': ['direct_hit', 'dart'],
+  '1F3AF': ['direct_hit', 'dart', 'bullseye'],
   // 🎱 pool 8 ball
   '1F3B1': ['billiards', '8ball'],
   // 🔮 crystal ball
@@ -3133,7 +3133,7 @@ export default {
   // 🧑 person
   '1F9D1': ['adult'],
   // 🧔 person: beard
-  '1F9D4': ['man_bearded'],
+  '1F9D4': ['person_bearded'],
   // 🧓 older person
   '1F9D3': ['older_adult'],
   // 🧕 woman with headscarf
@@ -3760,4 +3760,20 @@ export default {
   '26A7': ['transgender_symbol'],
   // 🏳️‍⚧️ transgender flag
   '1F3F3-FE0F-200D-26A7-FE0F': ['transgender_flag'],
+
+  // VERSION 13.1
+  // 😶‍🌫️ face in clouds
+  '1F636-200D-1F32B-FE0F': ['in_clouds'],
+  // 😮‍💨 face exhaling
+  '1F62E-200D-1F4A8': ['exhaling', 'exhale'],
+  // 😵‍💫 face with spiral eyes
+  '1F635-200D-1F4AB': ['dizzy_eyes'],
+  // ❤️‍🔥 heart on fire
+  '2764-FE0F-200D-1F525': ['heart_on_fire'],
+  // ❤️‍🩹 mending heart
+  '2764-FE0F-200D-1FA79': ['mending_heart'],
+  // 🧔‍♂️ man: beard
+  '1F9D4-200D-2642-FE0F': ['man_bearded'],
+  // 🧔‍♀️ woman: beard
+  '1F9D4-200D-2640-FE0F': ['woman_bearded'],
 };